requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
mcp>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.8.0
httpx>=0.27.0
pandas>=2.0.0
